Renamed hooks' files
